# openapi/auth.yaml
openapi: 3.0.3
info:
  title: Growfoundry Authentication API
  version: 2.0.0
  description: Authentication endpoints with separated auth and profile tables
paths:
...
  /api/auth/oauth/configs:
    get:
      summary: List all OAuth configurations
      description: Get all configured OAuth providers (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of OAuth configurations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OAuthConfig'
                  count:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only

    post:
      summary: Create OAuth configuration
      description: Create a new OAuth provider configuration (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
              properties:
                provider:
                  type: string
                  enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
                clientId:
                  type: string
                clientSecret:
                  type: string
                redirectUri:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                useSharedKey:
                  type: boolean
      responses:
        '200':
          description: OAuth configuration created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthConfig'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
  /api/auth/oauth/{provider}/config:
    get:
      summary: Get OAuth configuration for specific provider
      description: Get OAuth configuration including client secret (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
      responses:
        '200':
          description: OAuth configuration
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OAuthConfig'
                  - type: object
                    properties:
                      clientSecret:
                        type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
        '404':
          description: Configuration not found

    put:
      summary: Update OAuth configuration
      description: Update OAuth provider configuration (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                clientId:
                  type: string
                clientSecret:
                  type: string
                redirectUri:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                useSharedKey:
                  type: boolean
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthConfig'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
        '404':
          description: Configuration not found

    delete:
      summary: Delete OAuth configuration
      description: Delete OAuth provider configuration (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
      responses:
        '200':
          description: Configuration deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
        '404':
          description: Configuration not found
  /api/auth/oauth/{provider}:
    get:
      summary: Initiate OAuth flow (PKCE)
      description: |
        Generate OAuth authorization URL for any supported provider using PKCE flow.

        For mobile/desktop/server clients using PKCE:
        1. Generate code_verifier (random string, 43-128 chars)
        2. Generate code_challenge = Base64URL(SHA256(code_verifier))
        3. Call this endpoint with code_challenge
        4. After OAuth callback, use /api/auth/oauth/exchange with code_verifier
      tags:
        - Client
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: URL to redirect after authentication (receives growfoundry_code parameter)
        - name: code_challenge
          in: query
          required: false
          schema:
            type: string
          description: |
            PKCE code challenge for mobile/desktop/server clients.
            Generate using: Base64URL(SHA256(code_verifier))
            Required for secure token exchange in native apps or trusted server-side clients.
      responses:
        '200':
          description: OAuth authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  authUrl:
                    type: string
                    format: uri
                    description: URL to redirect user for OAuth provider login
        '400':
          description: Invalid request or provider not supported
        '500':
          description: OAuth not configured
...
  /api/auth/oauth/{provider}/callback:
    get:
      summary: Provider-specific OAuth callback (GET)
      description: |
        OAuth callback endpoint for provider-specific flows (most providers use GET).

        Response varies based on the original OAuth initiation:
        - With code_challenge (PKCE): Redirects with `growfoundry_code` for exchange endpoint
        - Without code_challenge (web): Redirects with `access_token` and sets httpOnly cookie
      tags:
        - Client
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
        - name: code
          in: query
          schema:
            type: string
          description: Authorization code from OAuth provider
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: JWT state with redirect URI and optional code_challenge
        - name: token
          in: query
          schema:
            type: string
          description: Direct ID token (for some providers)
      responses:
        '302':
          description: |
            Redirect to application.
            - PKCE flow: redirect_uri?growfoundry_code={code}&user_id={id}&email={email}&name={name}
            - Web flow: redirect_uri?access_token={token}&user_id={id}&email={email}&name={name}&csrf_token={csrf}
          headers:
            Location:
              schema:
                type: string
                format: uri
                description: Redirect URL with authentication parameters

    post:
      summary: Provider-specific OAuth callback (POST)
      description: |
        OAuth callback endpoint for providers that use POST (e.g., Apple with form_post response mode).

        Response varies based on the original OAuth initiation:
        - With code_challenge (PKCE): Redirects with `growfoundry_code` for exchange endpoint
        - Without code_challenge (web): Redirects with `access_token` and sets httpOnly cookie
      tags:
        - Client
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - state
              properties:
                code:
                  type: string
                  description: Authorization code from OAuth provider
                state:
                  type: string
                  description: JWT state with redirect URI and optional code_challenge
                id_token:
                  type: string
                  description: ID token (for Apple Sign In)
      responses:
        '302':
          description: |
            Redirect to application.
            - PKCE flow: redirect_uri?growfoundry_code={code}&user_id={id}&email={email}&name={name}
            - Web flow: redirect_uri?access_token={token}&user_id={id}&email={email}&name={name}&csrf_token={csrf}
          headers:
            Location:
              schema:
                type: string
                format: uri
                description: Redirect URL with authentication parameters
  /api/auth/oauth/custom/configs:
    get:
      summary: List all custom OAuth configurations
      description: Get all configured custom OAuth providers (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of custom OAuth configurations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomOAuthConfig'
                  count:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only

    post:
      summary: Create custom OAuth configuration
      description: Create a new custom OAuth provider configuration (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - key
                - name
                - discoveryEndpoint
                - clientId
                - clientSecret
              properties:
                key:
                  type: string
                name:
                  type: string
                discoveryEndpoint:
                  type: string
                  format: uri
                clientId:
                  type: string
                clientSecret:
                  type: string
      responses:
        '200':
          description: Custom OAuth configuration created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomOAuthConfig'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
        '409':
          description: Configuration already exists
  /api/auth/oauth/custom/{key}/config:
    get:
      summary: Get custom OAuth configuration
      description: Get a custom OAuth configuration including client secret (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Custom OAuth configuration
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CustomOAuthConfig'
                  - type: object
                    properties:
                      clientSecret:
                        type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
        '404':
          description: Configuration not found

    put:
      summary: Update custom OAuth configuration
      description: Update a custom OAuth provider configuration (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                discoveryEndpoint:
                  type: string
                  format: uri
                clientId:
                  type: string
                clientSecret:
                  type: string
      responses:
        '200':
          description: Custom OAuth configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomOAuthConfig'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
        '404':
          description: Configuration not found

    delete:
      summary: Delete custom OAuth configuration
      description: Delete a custom OAuth provider configuration (admin only)
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin only
        '404':
          description: Configuration not found
  /api/auth/oauth/custom/{key}:
    get:
      summary: Initiate custom OAuth flow (PKCE)
      description: Generate OAuth authorization URL for a configured custom OAuth provider using PKCE flow
      tags:
        - Client
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: URL to redirect after authentication (receives growfoundry_code parameter)
        - name: code_challenge
          in: query
          required: false
          schema:
            type: string
          description: PKCE code challenge for mobile/desktop/server clients
      responses:
        '200':
          description: OAuth authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  authUrl:
                    type: string
                    format: uri
        '400':
          description: Invalid request
        '404':
          description: Custom OAuth provider not found
  /api/auth/oauth/custom/{key}/callback:
    get:
      summary: Custom OAuth callback (GET)
      description: OAuth callback endpoint for custom OAuth providers
      tags:
        - Client
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from the custom OAuth provider
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: Signed OAuth state payload
      responses:
        '302':
          description: Redirect to application after custom OAuth completion
          headers:
            Location:
              schema:
                type: string
                format: uri
                description: Redirect URL with authentication parameters
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    OAuthConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [google, github, discord, linkedin, facebook, instagram, tiktok, apple, x, spotify, microsoft]
        clientId:
          type: string
          nullable: true
        redirectUri:
          type: string
          nullable: true
        scopes:
          type: array
          items:
            type: string
          nullable: true
        useSharedKey:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomOAuthConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        discoveryEndpoint:
          type: string
          format: uri
        clientId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...